Refactor position of Azure Devops validator
