+++ server/events/vcs/azuredevops/azuredevops_request_validator_test.go
@@ -1,4 +1,4 @@
-package server_test
+package azuredevops_test
 
 import (
 	"bytes"
@@ -6,14 +6,14 @@ import (
 	"testing"
 
 	. "github.com/petergtz/pegomock"
-	"github.com/runatlantis/atlantis/server"
+	"github.com/runatlantis/atlantis/server/events/vcs/azuredevops"
 	. "github.com/runatlantis/atlantis/testing"
 )
 
 func TestAzureDevopsValidate_WithBasicAuthErr(t *testing.T) {
 	t.Log("if the request does not have a valid basic auth user and password there is an error")
 	RegisterMockTestingT(t)
-	g := server.DefaultAzureDevopsRequestValidator{}
+	g := azuredevops.DefaultAzureDevopsRequestValidator{}
 	buf := bytes.NewBufferString("")
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -28,7 +28,7 @@ func TestAzureDevopsValidate_WithBasicAuthErr(t *testing.T) {
 func TestAzureDevopsValidate_WithBasicAuth(t *testing.T) {
 	t.Log("if the request has a valid basic auth user and password the payload is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultAzureDevopsRequestValidator{}
+	g := azuredevops.DefaultAzureDevopsRequestValidator{}
 	buf := bytes.NewBufferString(`{"yo":true}`)
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -43,7 +43,7 @@ func TestAzureDevopsValidate_WithBasicAuth(t *testing.T) {
 func TestAzureDevopsValidate_WithoutSecretInvalidContentType(t *testing.T) {
 	t.Log("if the request has an invalid content type an error is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultAzureDevopsRequestValidator{}
+	g := azuredevops.DefaultAzureDevopsRequestValidator{}
 	buf := bytes.NewBufferString("")
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -57,7 +57,7 @@ func TestAzureDevopsValidate_WithoutSecretInvalidContentType(t *testing.T) {
 func TestAzureDevopsValidate_WithoutSecretJSON(t *testing.T) {
 	t.Log("if the request is JSON the body is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultAzureDevopsRequestValidator{}
+	g := azuredevops.DefaultAzureDevopsRequestValidator{}
 	buf := bytes.NewBufferString(`{"yo":true}`)
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
